feat: port the game's sound: voices, music and 3D effects
hog_snd.zig ports sound_play and the voices it takes over by priority,
the fades, pausing, the positional sounds gathered over a frame, the music
with its loop table, and the 3D voices kept placed from the camera.
sound3d.zig ports the 3D effects: their definitions, generated from the
executable by tablegen, the voice classes, the player's engine and
afterburner, and the fighters heard flying past.

+++ src/engine/mss.zig
@@ -20,14 +20,14 @@ pub const Status = voice.Status;
 pub const Vector = positional.Vector;
 
 /// A sample (`HSAMPLE`): a sound played as it is, with a volume and a pan.
-pub const Sample = enum(u8) { _ };
+pub const Sample = enum(u32) { _ };
 
 /// A 3D sample (`H3DSAMPLE`): a sound placed around the listener.
-pub const Sample3D = enum(u8) { _ };
+pub const Sample3D = enum(u32) { _ };
 
 /// A stream (`HSTREAM`): a long sound, such as a piece of music, that plays and loops as a sample
 /// does, from a file of its own.
-pub const Stream = enum(u8) { _ };
+pub const Stream = enum(u32) { _ };
 
 /// What keeps the platform's mixing thread out while the game changes the driver.
 pub const Lock = struct {
